+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## next
 
+- Introduced basic support for function arguments in the syntax, allowing `$arg => expr` and `($a, $b) => expr` forms to be supported
 - Changed the behavior of `this` in custom methods and assertions defined as functions to include a `context` reference to the query context. Additionally, introduced `this.method(name, ...args)` and `this.assertion(name, ...args)` methods to call a custom method or use an assertion.
 - Added support for special variables `$$` and `#` in custom methods and assertions defined as strings (Jora queries).
 - Updated the pipeline operator grammar to allow starting a query with it, e.g. `| expr` is now a valid query